// src/js/types/bookmark-pdf-type.ts
export type BookmarkTarget =
  | { kind: 'uri'; url: string; newWindow?: boolean }
  | { kind: 'destination'; type: string; args: (number | null)[] };